\name{f.dir.create}
\alias{f.dir.create}
\title{
Create Output Path for a Folder
}
\description{
Function to create output path. Users will be asked to choose directory and input the name of a folder. No arguments needs
}
\usage{
f.dir.create()
}
\value{value returning list of path of directory and name of folder}

