\name{GeneBook-package}
\alias{GeneBook-package}
\alias{GeneBook}
\docType{package}
\title{
\packageTitle{GeneBook}
}
\description{
\packageDescription{GeneBook}
}
\details{
GeneBook package applies to the strings of genes. It can convert gene ID and its alias to a consistent symbol and provide information about how accurate the conversion is. Furthermore, it can tell detailed information about genes. fThis package relies on our database which includes data from GeneCards Human Gene Database.

GeneCard_main: The main function to export converted genes and genes description after import a file including genes interested

GeneCard_ID_Convert: Convert genes to consistent symbols

GeneCard_Symbol_Details: Function to get the detailed information of gene
}
\author{
\packageAuthor{GeneBook}

Maintainer: \packageMaintainer{GeneBook}
}

\keyword{ package }




