\name{genecard_description_summary}
\alias{genecard_description_summary}
\docType{data}
\title{Gene Description}
\description{
A dataset containing the consistent gene symbol, type  and description
}
\usage{genecard_description_summary}
\format{A dataframe contains 59,918 observations and 7 columns}
\source{Human Gene Database}
\keyword{datasets}

