\name{GeneCard_main}
\alias{GeneCard_main}
\title{
MAIN FUNCTION FOR THE PACKAGE for Mac Users
}
\description{
It allows users to choose the file which includes a column of genes interested and create a folder for output. File can be CSV or RDA. This function will export two tables of the converted gene symbols and gene description. No arguments needs
}
\usage{
GeneCard_main()
}
\value{
Value returning is a folder with path and name assigned by users. Two TXT files are under the folder including a file of converted genes and a file of genes description
}
\details{
This function depends on the function of choose.dir() and f.dir.creat()
}



