\name{get_os}
\alias{get_os}
\title{Get Information about Operating System
}
\description{
Generic function tells R operating system information. No argument needs
}
\usage{
get_os()
}
\value{value returning is a string providing the Information about users' Operating System }
\references{
http://conjugateprior.org/2015/06/identifying-the-os-from-r/
}
\author{
Will Lowe
}
\note{
Function get_os()is derived from the work of others
}
\examples{
get_os()
}
