\name{dir.choose}
\alias{dir.choose}
\title{Function to choose directory for both Mac users }
\description{
This function replaces function choose.dir()
}
\usage{dir.choose()}
\references{
https://grokbase.com/t/r/r-sig-mac/12bxhv5xcz/equivalent-of-choose-dir
}
\author{
Hans Bibiko
}
\note{
Function dir.choose() is derived from the work of others
}
