\name{GeneCard_main}
\alias{GeneCard_main}
\title{
MAIN FUNCTION FOR THE PACKAGE for Mac Users
}
\description{
Import the file which includes genes interested. This function will export two tables of the converted gene symbols and gene description
}
\usage{
GeneCard_main()
}
\details{
This function depends on the function of choose.dir() and f.dir.creat()
}



