\name{get_os}
\alias{get_os}
\title{Get Information about Operating System
}
\description{
Generic function tells R operating system information
}
\usage{
get_os()
}
\references{
http://conjugateprior.org/2015/06/identifying-the-os-from-r/
}
\author{
Will
}
\note{
Function get_os()is derived from the work of others
}
